LIME-4: * fixed button folder changed
