+++ frontend/src/bundles/common/helpers/get-valid-class-names.ts
